+++ Packages/Vanilla JNode/Runtime/Samples/Earth Sample/ContinentEditor.cs
@@ -1,0 +1,13 @@
+using Vanilla.JNode.Samples;
+
+namespace Vanilla.JNode
+{
+
+	public class ContinentEditor : JNodeEditor<Continent>
+	{
+
+		protected override string GetName() => data.Name;
+
+	}
+	
+}