+++ Packages/Vanilla JNode/Runtime/Samples/Earth Sample/CountryEditor.cs
@@ -1,0 +1,13 @@
+using Vanilla.JNode.Samples;
+
+namespace Vanilla.JNode
+{
+
+    public class CountryEditor : JNodeEditor<Country>
+    {
+
+        
+        protected override string GetName() => data.Name;
+
+    }
+}