+++ Packages/Vanilla Normal/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 21-04-2022
+
+### Removed
+- EasingNormal has been deprecated.
+- Vanilla Easing and Type Menu are no longer dependencies.
+
 ## [1.1.0] - 09-03-2022
 
 ### Changed